deploy: handle first-time setup better

+++ rhel7/deploy.sh
@@ -50,15 +50,22 @@ sudo systemctl stop httpd || true
 sudo systemctl stop nca-httpd || true
 sudo systemctl stop nca-workers || true
 
-echo Removing the old stuff
+src=$(pwd)
 ncadir=/usr/local/nca
-tmpdir=/tmp/nca-$(date +"%s")
-sudo mv $ncadir $tmpdir
-sudo mkdir $ncadir
-sudo mv $tmpdir/settings* $ncadir/
-sudo find $ncadir/ -mindepth 1 -maxdepth 1 -type f -not -name "settings*" -exec rm -f {} \;
-sudo rm $ncadir/templates/ -rf
-sudo rm $ncadir/static/ -rf
+if [[ -d $ncadir ]]; then
+  echo Removing the old stuff
+  tmpdir=/tmp/nca-$(date +"%s")
+  sudo mv $ncadir $tmpdir
+  sudo mkdir $ncadir
+  sudo mv $tmpdir/settings* $ncadir/
+  sudo find $ncadir/ -mindepth 1 -maxdepth 1 -type f -not -name "settings*" -exec rm -f {} \;
+  sudo rm $ncadir/templates/ -rf
+  sudo rm $ncadir/static/ -rf
+else
+  echo First-time install detected: edit $ncadir/settings
+  sudo mkdir $ncadir
+  sudo cp $src/settings-example $ncadir/settings
+fi
 
 echo Removing the cache
 sudo rm /tmp/nca/finder.cache -f
@@ -67,7 +74,6 @@ echo Migrating the database
 goose --env production up
 
 echo Copying in the new stuff
-src=$(pwd)
 sudo cp $src/bin/server $ncadir/
 sudo cp $src/bin/run-jobs $ncadir/
 sudo cp $src/bin/queue-batches $ncadir/
@@ -78,6 +84,8 @@ sudo cp -r $src/templates/ $ncadir/
 sudo cp -r $src/static/ $ncadir/
 
 echo Doing a daemon reload and starting the service
+sudo systemctl enable $ncadir/nca-httpd.service
+sudo systemctl enable $ncadir/nca-workers.service
 sudo systemctl daemon-reload
 sudo systemctl start nca-workers
 sudo systemctl start nca-httpd
